blog, refined and tested
Tidied up the base code I worked up for the first commit. I think I did a pretty darn decent job of matching the example, with a few tiny bits of personal flair for Extra Flavor...! (color gradients, centered the nav links, etc.) Struggled to get the <button> code working, unfortunately, so it looks pretty, but it's... not exactly functional. Still! I'm pretty pleased with the outcome all the same!